requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==5.1.0
markdown2==2.4.12
mysqlclient==2.2.1
openai==1.8.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
readtime==3.0.0
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
